add types to components & take out NameForm
