+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\dater-test\\dater-test-game\\src\\index.tsx":"1","D:\\dater-test\\dater-test-game\\src\\App.tsx":"2","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx":"3"},{"size":352,"mtime":1608019368940,"results":"4","hashOfConfig":"5"},{"size":327,"mtime":1608033361157,"results":"6","hashOfConfig":"5"},{"size":536,"mtime":1608033830839,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"gvxkxr",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\dater-test\\dater-test-game\\src\\index.tsx",[],["15","16"],"D:\\dater-test\\dater-test-game\\src\\App.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\dater-test\\dater-test-game\\src\\index.tsx":"1","D:\\dater-test\\dater-test-game\\src\\App.tsx":"2","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx":"3","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx":"4"},{"size":352,"mtime":1608019368940,"results":"5","hashOfConfig":"6"},{"size":363,"mtime":1608034668229,"results":"7","hashOfConfig":"6"},{"size":161,"mtime":1608034992972,"results":"8","hashOfConfig":"6"},{"size":757,"mtime":1608034980748,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"gvxkxr",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\dater-test\\dater-test-game\\src\\index.tsx",[],["19","20"],"D:\\dater-test\\dater-test-game\\src\\App.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
