+++ src/pages/StartScreen/index.tsx
@@ -1,19 +1,8 @@
-import {Button, List, ListInput, Popup} from "framework7-react"
+import React from "react"
+import NameForm from "./NameForm"
 
-const popupCustomStyle = {
-    maxHeight: "180px",
-    marginTop: "-120px",
-}
-
-const StartScreen = () => {
-    return (
-        <Popup closeByBackdropClick={false} opened={true} style={popupCustomStyle}>
-            <List inlineLabels>
-                <ListInput type="text" label="Name" placeholder="Your name" autofocus clearButton />
-            </List>
-            <Button>Confirm Name</Button>
-        </Popup>
-    )
+const StartScreen: React.FC = () => {
+    return <NameForm />
 }
 
 export default StartScreen
